+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Add `show_locator_types` attribute to `FindingGroup` to filter locators shown in the report.
 - Add support for Python 3.14 to Tox configuration.
 - Add more tests for `VersionConfig`.
 
@@ -26,7 +27,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Update pypdf requirement from ~=6.0.0 to ~=6.1.0
 - Update cryptography requirement from ~=45.0.2 to ~=46.0.1
 - Update pydantic-settings requirement from ~=2.10.0 to ~=2.11.0
-
 
 ## [0.3.7] - 2025-09-09